LayoutTests/imported/w3c/web-platform-tests/webrtc-encoded-transform/encoded-frame-constructors.https.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <meta name="timeout" content="long">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id="video1" autoplay></video>
        <video id="video2" autoplay></video>
        <script src ="routines.js"></script>
        <script>
function validateFrame(frame1, frame2, excludedProperties = [])
{
    assert_equals(frame1.timestamp, frame2.timestamp);
    if (frame1.data) {
        assert_array_equals(new Uint8Array(frame1.data), new Uint8Array(frame2.data));
    } else
        assert_equals(frame2.data, null);

    const metadata1 = frame1.getMetadata();
    const metadata2 = frame1.getMetadata();

    assert_array_equals(Object.getOwnPropertyNames(metadata1), Object.getOwnPropertyNames(metadata2)); 
    for (const property in metadata1) {
       if (excludedProperties.indexOf(property) !== -1)
           continue;
       if (property === "contributingSources" || property === "dependencies")
           assert_array_equals(metadata1[property], metadata2[property], property);
       else
           assert_equals(metadata1[property], metadata2[property], property);
    }
}

function validateAudioFrame(frame1, frame2, excludedProperties)
{
    assert_equals(frame1.type, frame2.type, excludedProperties);
}

function validateVideoFrame(frame1, frame2, excludedProperties)
{
    assert_equals(frame1.type, frame2.type);
    validateFrame(frame1, frame2, excludedProperties);
}

async function getAndValidateAudioFrame(t, port, testName)
{
    port.postMessage("getFrame");
    const frame = await new Promise((resolve, reject) => {
        port.onmessage = e => resolve(e.data);
        t.step_timeout(() => reject("getFrame timed out"), 1000);
    });

    assert_true(frame instanceof RTCEncodedAudioFrame);

    let clone = new RTCEncodedAudioFrame(frame);
    test(() => {
        validateAudioFrame(frame, clone);
    }, testName + " - clone");

    clone = new RTCEncodedAudioFrame(frame, { metadata : { rtpTimestamp : 10 } });
    test(() => {
        assert_equals(clone.getMetadata().rtpTimestamp, 10);
        validateAudioFrame(frame, clone, ["rtpTimestamp"]);
    }, testName + " - clone - rtpTimestamp");
}

async function getAndValidateVideoFrame(t, port, testName)
{
    port.postMessage("getFrame");
    const frame = await new Promise((resolve, reject) => {
        port.onmessage = e => resolve(e.data);
        t.step_timeout(() => reject("getFrame timed out"), 1000);
    });

    assert_true(frame instanceof RTCEncodedVideoFrame);

    let clone = new RTCEncodedVideoFrame(frame);

    test(() => {
        validateVideoFrame(frame, clone);
    }, testName + " - clone");

    clone = new RTCEncodedVideoFrame(frame, { metadata : { rtpTimestamp : 10 } });
    test(() => {
        assert_equals(clone.getMetadata().rtpTimestamp, 10);
        validateAudioFrame(frame, clone, ["rtpTimestamp"]);
    }, testName + " - clone - rtpTimestamp");

    clone = new RTCEncodedVideoFrame(frame, { metadata : { synchronizationSource : 10 } });
    test(() => {
        assert_equals(clone.getMetadata().synchronizationSource, 10);
        validateAudioFrame(frame, clone, ["synchronizationSource"]);
    }, testName + " - clone - synchronizationSource");

    clone = new RTCEncodedVideoFrame(frame, { metadata : { frameId : 10 } });
    test(() => {
        assert_equals(clone.getMetadata().frameId, 10);
        validateAudioFrame(frame, clone, ["frameId"]);
    }, testName + " - clone - frameId");
}

async function getInboundRTPStats(receiver)
{
    const report = await receiver.getStats();
    var stats;
    report.forEach((statItem) => {
        if (statItem.type === "inbound-rtp") {
            stats = statItem;
        }
    });
    return stats;
}

async function getDecodedFrameCount(receiver)
{
    const stats = await getInboundRTPStats(receiver);
    return stats ? stats.framesDecoded : 0;
}

async function checkDecodedFrameCountIsIncreasing(test, receiver, testName)
{
    const frameCount = await getDecodedFrameCount(receiver);
    let counter = 0;
    do {
        await waitFor(test, 200); 
        const newFrameCount = await getDecodedFrameCount(receiver);
        if (newFrameCount !== frameCount)
            break;
    } while (++counter < 20);

    assert_less_than(counter, 20, "checkDecodedFrameCountIsIncreasing " + (testName ? testName : ""));
}

async function checkDecodedFrameCountIsFreezing(test, receiver, testName)
{
    let frameCount = await getDecodedFrameCount(receiver);
    let counter = 0;
    do {
        await waitFor(test, 200); 
        const newFrameCount = await getDecodedFrameCount(receiver);
        if (newFrameCount === frameCount)
            break;
        frameCount = newFrameCount;
    } while (++counter < 20);

    assert_less_than(counter, 20, "checkDecodedFrameCountIsFreezing " + (testName ? testName : ""));
}

promise_test(async (test) => {
    const {sender, receiver, senderPc, receiverPc} = await createConnectionWithTransform(test, 'encoded-frame-worker.js', {video: true});
    return getAndValidateVideoFrame(test, sender.transform.port, "Sender RTCEncodedVideoFrame");
}, "Create sender RTCEncodedVideoFrame");

promise_test(async (test) => {
    const {sender, receiver, senderPc, receiverPc} = await createConnectionWithTransform(test, 'encoded-frame-worker.js', {video: true});
    return getAndValidateVideoFrame(test, receiver.transform.port, "Receiver RTCEncodedVideoFrame");
}, "Create receiver RTCEncodedVideoFrame");

promise_test(async (test) => {
    const {sender, receiver, senderPc, receiverPc} = await createConnectionWithTransform(test, 'encoded-frame-worker.js', {audio: true});
    return getAndValidateAudioFrame(test, sender.transform.port, "Sender RTCEncodedAudioFrame");
}, "Create sender RTCEncodedAudioFrame");

promise_test(async (test) => {
    const {sender, receiver, senderPc, receiverPc} = await createConnectionWithTransform(test, 'encoded-frame-worker.js', {audio: true});
    return getAndValidateAudioFrame(test, receiver.transform.port, "Receiver RTCEncodedAudioFrame");
}, "Create receiver RTCEncodedAudioFrame");
       </script>
    </body>
</html>
